rename: const/categories to category
